LocalStorage
This is the sample local storage project. It provides a local storage class that allows you to pass in a property name and object which is stored in a dictionary encrypted. The values can be retreived from the local storage and the storage is saved to file and loaded from file.